handling last day of month bug

+++ srvSheet.R
@@ -160,7 +160,7 @@ hot_dat2DF <- function(data, repoStruct, orderDecreasing){
                        },
                        timestamp = {
                                DF[, fields[i]] <- 
-                                       as.integer(DF[, fields[i]])
+                                       as.double(DF[, fields[i]])
                                
                        },
                        boolean = {
